CurvePoint delete refactor

+++ src/main/java/com/nnk/springboot/service/CurvePointService.java
@@ -59,8 +59,13 @@ public class CurvePointService {
 
     }
 
-    public void deleteCurvePoint(CurvePoint curvePoint){
-        curvePointRepository.delete(curvePoint);
-        log.info("CurvePoint deleted");
+    public void deleteCurvePoint(int id){
+        CurvePoint curvePoint = curvePointRepository.findCurvePointById(id);
+        if (curvePoint != null){
+            curvePointRepository.delete(curvePoint);
+            log.info("CurvePoint successfully deleted.");
+        } else {
+            log.error("Deletion failure");
+        }
     }
 }
